backend/pyproject.toml
[project]
name = "competition-judge-system"
version = "0.1.0"
description = "中国大学生创新大赛AI评委系统后端"
requires-python = ">=3.13"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "supabase>=2.15.0",
  "python-multipart>=0.0.18",
  "python-pptx>=1.0.0",
  "Pillow>=11.0.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0",
  "python-docx>=1.1.0",
  "openpyxl>=3.1.0",
  "PyPDF2>=3.0.0",
  "pytest>=8.0",
  "pytest-asyncio>=1.0",
  "hypothesis>=6.100.0",
  "reportlab>=4.4.10",
  "getstream[webrtc,telemetry]>=2.5.21,<3.0.0",
  "vision-agents>=0.2.8",
  "vision-agents-plugins-getstream",
  "vision-agents-plugins-qwen",
]
